+++ spec/truffle/interop/fixtures/big5_magic.rb
@@ -1,0 +1,2 @@
+# encoding: big5
+__ENCODING__.name