+++ spec/truffle/interop/fixtures/no_magic.rb
@@ -1,0 +1,1 @@
+__ENCODING__.name