+++ spec/truffle/interop/fixtures/usascii_magic.rb
@@ -1,0 +1,2 @@
+# encoding: us-ascii
+__ENCODING__.name